pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nifi-mcp-server"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Apache NiFi"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "kimcaprio", email = "kimcaprio@github.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "streamlit>=1.31.0",
    "loguru>=0.7.2",
    "pytest>=7.4.3",
    "python-json-logger>=2.0.7",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
    "pytest-cov",
]

[project.urls]
"Homepage" = "https://github.com/kimcaprio/NIFI_MCP_SERVER"
"Bug Tracker" = "https://github.com/kimcaprio/NIFI_MCP_SERVER/issues"

[tool.hatch.build.targets.wheel]
packages = ["nifi_mcp_server", "nifi_chat_ui"]
